refactor(hooks): move use-mobile hook to useIsMobile and update import paths
