+++ frontend/src/components/ui/sidebar.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { VariantProps, cva } from 'class-variance-authority';
-import { useIsMobile } from '@/hooks/use-mobile';
+import { useIsMobile } from '@/app/hooks/useIsMobile';
 import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
